+++ src/main/java/com/kikis/ptdyeplus/kubejs/BaseBindings.java
@@ -1,0 +1,5 @@
+package com.kikis.ptdyeplus.kubejs;
+
+public interface BaseBindings {
+    Utils utils = new Utils();
+}