+++ changelog.md
@@ -1,3 +1,5 @@
+1.2.4 - Remove getOptions function, add getKey function<br>
+1.2.3 - Exposes function to KubeJS for getting options.txt file<br>
 1.2.2 - Compiled full list of ponder-able blocks<br>
 1.2.1 - Adds 'Press x to ponder' on Jade tooltip to ponder-able blocks<br>
 1.2.0 - Adds key bind to open create ponder on what block the player is looking at<br>